feat: Moves search button into a slot
